refactor(Slots): SlotsGame folder moved to Slots/components
