+++ components/Slots/components/SlotsGame/SlotsGame.tsx
