+++ components/Slots/components/SlotsGame/components/ProductCard/components/CardContent/index.ts
