+++ components/Slots/components/SlotsGame/components/ProductCard/components/CardDynamicGradient/index.ts
