+++ components/Slots/components/SlotsGame/components/ProductCard/components/CardImage/index.ts
@@ -1,0 +1,1 @@
+export { default as CardImage } from "./CardImage";
