+++ components/Slots/components/SlotsGame/components/ProductCard/components/TierBadge/components/BadgeContent/BadgeContent.tsx
