+++ components/Slots/components/SlotsGame/components/ProductCard/components/TierBadge/components/BadgeDynamicGradiet/BadgeDynamicGradient.tsx
