+++ components/Slots/components/SlotsGame/components/ProductCard/components/TierBadge/components/BadgeDynamicGradiet/index.ts
