+++ components/Slots/components/SlotsGame/components/ProductCard/components/TierBadge/index.ts
