+++ components/Slots/components/SlotsGame/components/ProductCard/utils/index.ts
