+++ components/Slots/components/SlotsGame/components/ProductSlots/index.ts
