+++ components/Slots/components/SlotsGame/components/SpinButton/index.ts
