+++ components/Slots/components/SlotsGame/index.ts
