+++ components/Slots/components/SlotsTitle/SlotsTitle.tsx
@@ -1,0 +1,8 @@
+const SlotsTitle = () => {
+  return (
+    <h2 className="md:text-4xl text-xl font-bold mb-4 mt-2 ml-8 text-slate-100">
+      Roleta de Ofertas
+    </h2>
+  );
+};
+export default SlotsTitle;
