+++ components/Slots/components/SlotsTitle/index.ts
@@ -1,0 +1,1 @@
+export { default as SlotsTitle } from "./SlotsTitle";
